+++ mono/utils/strenc.c
@@ -41,13 +41,13 @@ static const char trailingBytesForUTF8[256] = {
  * of bytes in the returned string, not including the terminator.
  */
 gunichar2 *
-mono_unicode_from_external (const gchar *in, gsize *bytes)
+mono_unicode_from_external (const gchar *in, guint32 *bytes)
 {
 	gchar *res=NULL;
 	gchar **encodings;
 	gchar *encoding_list;
 	int i;
-	glong lbytes;
+	guint32 lbytes;
 	
 	if(in==NULL) {
 		return(NULL);
